Refactor file-summary header
